Extract common server-management code to parent route
